Cleanup from lucene version change
I keep hedging on whether to keep a self-standing executable jar checked
in.  I don't really want it; it's there for convenience for users
who want to use the tool but don't want to do maven builds.  For now, I
just rebuilt the jar.

For lucene version changes, master should probably track lucene's latest
stable release, and we can have branches for older versions if needed.
For my purposes, looks like building with lucene 4.5 (and using
LUCENE_45 in the analyzer API) still works with indexes I've built with
lucene 4.3, so might as well just go with the latest here.